refactor: introduce new authentication method API
This is the new typestate-based auth method API that provides open-ended
extensibilty (i.e. a library user can add their own methods, using for
example the 80h-FEh private ID range specified by the RFC).

It provides a lot of safety, except for the possibility for two methods
to declare the same ID. Sadly, functions in traits cannot be const,
or I would've written a static assert at least in the enum macro..