fix(server): don't stop UDP proxying on single packet errors
We're not supposed to stop upon encountering an error, we're supposed
to drop the packet and go on. Let's split the inner parts of the loops
into functions that return Result and match on that to just debug log.

While here, reduce ridiculous 64K buffer sizes and simplify some silly code.